requirements.txt
requests
beautifulsoup4
selenium
google-generativeai
webdriver-manager
Jinja2
